audio_solutions.py
"""
AstrBot 音频文件发送方案探索

根据错误信息：'WebChatMessageEvent' object has no attribute 'file_result'
说明WebChatMessageEvent不直接支持file_result()方法。

可能的解决方案：

方案1: 使用MessageEventResult构建自定义消息
-----------------------------------------------
from astrbot.api.event import MessageEventResult

# 可能的用法（需要查阅AstrBot文档确认）:
result = MessageEventResult()
result.use_custom_result("voice", {"path": voice_path})
yield result

方案2: 转换音频格式
-----------------------------------------------
某些平台（如QQ、微信）对语音格式有特殊要求：
- QQ: 需要silk/amr格式
- 微信: 需要silk格式
- Discord: 支持mp3

可以使用ffmpeg或pydub转换格式:
```python
from pydub import AudioSegment

# MP3转WAV
audio = AudioSegment.from_mp3(voice_path)
wav_path = voice_path.replace('.mp3', '.wav')
audio.export(wav_path, format='wav')

# 或使用ffmpeg
import subprocess
subprocess.run(['ffmpeg', '-i', voice_path, '-ar', '8000', '-ac', '1', output_path])
```

方案3: Base64编码嵌入
-----------------------------------------------
将音频转换为base64字符串，通过文本消息发送数据URI:
```python
import base64

with open(voice_path, 'rb') as f:
    audio_data = base64.b64encode(f.read()).decode()
    data_uri = f"data:audio/mp3;base64,{audio_data}"
    yield event.plain_result(f"[语音消息] {data_uri[:100]}...")
```

方案4: 提供HTTP下载链接
-----------------------------------------------
启动一个简单的HTTP服务器，提供语音文件下载：
```python
from http.server import HTTPServer, SimpleHTTPRequestHandler
import threading

# 在插件初始化时启动HTTP服务器
def start_http_server(port=8000):
    server = HTTPServer(('0.0.0.0', port), SimpleHTTPRequestHandler)
    thread = threading.Thread(target=server.serve_forever)
    thread.daemon = True
    thread.start()

# 发送下载链接
voice_url = f"http://localhost:8000/{voice_path}"
yield event.plain_result(f"🔊 语音文件: {voice_url}")
```

方案5: 使用平台特定API（推荐）
-----------------------------------------------
检查AstrBot的平台适配层，使用平台原生的语音发送API:
```python
# 检查事件类型
if hasattr(event, 'platform'):
    if event.platform == 'qq':
        # 使用QQ的语音发送API
        pass
    elif event.platform == 'wechat':
        # 使用微信的语音发送API
        pass

# 或检查是否有平台特定的方法
if hasattr(event, 'send_voice'):
    yield event.send_voice(voice_path)
elif hasattr(event, 'send_record'):
    yield event.send_record(voice_path)
```

当前实现（方案6）: 提供本地文件路径
-----------------------------------------------
最简单的方案是告知用户语音文件的本地路径:
```python
yield event.plain_result(f"🔊 生日语音已下载\\n文件位置: {voice_path}")
```

用户可以手动播放或通过其他方式分享。

建议测试顺序:
1. 检查AstrBot文档，查找官方推荐的音频发送方法
2. 尝试MessageEventResult的自定义类型
3. 检查event对象是否有其他发送方法（send_voice, send_record等）
4. 如果都不支持，使用HTTP服务器提供下载链接
"""

if __name__ == "__main__":
    print(__doc__)
